Remove createFilterMenuAppender.ts
