--- src/dmm/books/add-filter-button-on-bookmark-page/createFilterMenuAppender.ts
@@ -1,18 +1,0 @@
-// TODO: リファクタリング
-
-import { createFilterMenu } from './createFilterMenu'
-import { FilterType } from './FilterType'
-import { findMenu } from './findMenu'
-
-export const createFilterMenuAppender = (filterType: FilterType): (main: Element) => void => {
-  return (main) => {
-    const menu = findMenu(main)
-
-    if (menu === undefined) {
-      throw new Error('Missing menu.')
-    }
-
-    const filterMenu = createFilterMenu(filterType)
-    menu.append(filterMenu)
-  }
-}
